+++ database/migrations/calculator_loan_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
 
             $table->integer('type')->nullable()->comment('대출 구분 - 1: 원금균등분할, 2: 원리금균등분할, 3: 만기일시');
 
-            $table->integer('loan_price')->nullable()->comment('대출 원금');
+            $table->bigInteger('loan_price')->nullable()->comment('대출 원금');
             $table->double('loan_rate', 10, 2)->nullable()->comment('이자율 소수점 2자리까지');
             $table->integer('loan_month')->nullable()->comment('대출 기간 (개월)');
             $table->integer('holding_month')->nullable()->comment('거치 기간 (개월)');
